Disable optimistic flag on updates
The implementation of this feature can cause data loss
if database compaction finishes immediately before an
optimistic updates gain access to the updater.
For e.g., the request just wrote the document body
to the file, in the meanwhile database compaction
finished, the old file (where the updater wrote the
document body to) was deleted and we write #doc_info
and #full_doc_info records into the btrees referring
to the offset, in the old file, where the document
body was written to.

Change-Id: Id026d8bc12397e9e93d65b5ae23f0736e3b9688f
Reviewed-on: http://review.couchbase.org/8325
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>