Got rid of the mc_daemon workers in favor of a supervisor.
Every time I see trap_exit, I imagine a bug is coming shortly after.
I think this is closer to the structure that's actually wanted here.
A supervisor launches temporary children and maintains the control of
them.

I do still ask for notification of the death of these processes, so I
can reissue batches, but there's no chance of a proper exit not
shutting us all the way down.

Change-Id: Icbb4c44a9efb77ff1996106e69c5a96d344583a2
Reviewed-on: http://review.couchbase.org/8661
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>