MB-4163 Maintain the number of workers through monitor refs array
We found that the batch supervisor sometimes sends "DOWN" message
to the monitoring process even before a given worker is actually
removed from the supervisor's child list. Consequently, this
erlang messaging behavior caused the race condition in the
monitoring process. To resolve this issue, this change maintains
the number of batch workers by keeping track of the array of
monitor references.

Change-Id: I835261bbd26e4412d345dc210d99eb12ee5cfd45
Reviewed-on: http://review.couchbase.org/8882
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>