made vbuckets creation quicker. Related to MB-4466
Shutting down buckets sometimes takes lots of time which is causing
node join to time out. This is happening when ep-engine's mc-kvstore
is busy in snapshot_vbucket_states command. That command takes a lot
of time to complete when it has to create lots of vbucket databases in
mccouch. This commit speeds it up by creating vbuckets databases in
parallel.

Change-Id: I6c6d110cb4691d526261cdea492b09066c1ce21d
Reviewed-on: http://review.couchbase.org/11246
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>