Persist/Load the per vbucket MaxDeletedSeqno values
The max deleted seqno values are initialized to _local/vbstate as part
set_vbucket_state() function the first time vbuckets are created.

The persisted values are returned as part of get_state() which is
used during warmup to build the initial state.

A separate change set will add the code to incrementally update the
seqno values as delete mutations are received and persisted to Couch.

Change-Id: Ic63c77d4fa3724382ca05def3d5ca1681251e0a7
Reviewed-on: http://review.couchbase.org/11903
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>